requirement.txt
langchain
validators

youtube_transcript_api
unstructured
pytube
playwright
selenium
streamlit

langchain==0.1.16
langchain_community==0.0.33
langchain_openai==0.0.8
Pypdf
Pinecone
load_dotenv
giskard
giskard[llm]
openai==1.12.0
tiktoken==0.6.0
mlflow==2.12.1
faiss-cpu 
starlette
fastapi
watchfiles
slowapi
gunicorn
waitress
evaluate
nltk
rouge-score
transformers
torch
textstat
dagshub
langchain
huggingface_hub
pinecone
langchain_pinecone
sentence_transformers
chromadb
faiss-cpu
